pyproject.toml
# PyProject configuration for VMSCA experiment

[project]
name = "vmsca-experiment"
version = "0.1.0"
description = "Verified Metacognitive Self-Consistency with Adaptive Stopping"
requires-python = ">=3.10"

dependencies = [
    # Core experiment framework
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    
    # Deep learning and transformers
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    
    # Datasets
    "datasets>=2.16.0",
    
    # Hyperparameter optimization
    "optuna>=3.5.0",
    
    # Logging and tracking
    "wandb>=0.16.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    
    # Utilities
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"
